% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Homework.R
\name{count5test}
\alias{count5test}
\title{The power of the CountFive test in Exercise 6.8 in Homework4}
\usage{
count5test(x, y)
}
\arguments{
\item{x}{the sample named X}

\item{y}{the sample named Y}
}
\value{
the value 1 or 0: reject or accept H0
}
\description{
The two sample “Count Five” test for equality of variance introduced by McGrath and Yeh [193] counts the number of extreme points of each sample relative to the range of the other sample.
}
\examples{
 
\dontrun{
X=rnorm(1000,0,2)
Y=rnorm(1000,0,1)
count5test(X,Y)
}
}
