% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Homework.R
\name{f1}
\alias{f1}
\title{The important function(1) for monte carlo in Exercise5.13, Homework3}
\usage{
f1(x)
}
\arguments{
\item{x}{the value of the independent variable}
}
\value{
the value of the function at \code{x}
}
\description{
a important function help us to use the Monte Carlo method to integrate the goal density
}
