% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Homework.R
\name{f_inv}
\alias{f_inv}
\title{Pareto(a, b)^-1 in the Exercises 3.3, Homework1}
\usage{
f_inv(x, a, b)
}
\arguments{
\item{x}{the value of the independent variable}

\item{a}{the exponent part}

\item{b}{the Molecular part}
}
\value{
the value when inverse distribution with parameter\code{(a,b)} at \code{x}.
}
\description{
The inverse distribution for Pareto with parameter(a,b)
}
\examples{
\dontrun{
n <- 1000
u <- runif(n)
x <- f_inv(u,2,2)
hist(x, density = 3,breaks=50, col="red",prob = TRUE)
}
}
