% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Homework.R
\name{f_j}
\alias{f_j}
\title{The important function in the monte carlo in Exercise5.15, Homework3}
\usage{
f_j(x, j, k)
}
\arguments{
\item{x}{the value of the independent variable}

\item{j}{the order of the piecewise}

\item{k}{the total number of the piecewise}
}
\value{
the value of the function at \code{x}
}
\description{
a important function help us to use the Monte Carlo piecewise method to integrate the goal density
}
