% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Homework.R
\name{F_Mv}
\alias{F_Mv}
\title{lapply variant, in Ex4, Homework8}
\usage{
F_Mv(f, n, type, ...)
}
\arguments{
\item{f}{The function that operates on data}

\item{n}{The dimension of data}

\item{type}{the data type: numeric, character, complex, logical}

\item{...}{the input}
}
\value{
the data after the operation
}
\description{
The function is a combination of Map() and vapply() to create an lapply() variant that iterates in parallel over all of its inputs and stores its outputs in a vector (or a matrix).
}
