% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Homework.R
\name{f_Pareto}
\alias{f_Pareto}
\title{Pareto(a, b) in the Exercises 3.3, Homework1}
\usage{
f_Pareto(x, a, b)
}
\arguments{
\item{x}{the value of the independent variable}

\item{a}{the exponent part}

\item{b}{the Molecular part}
}
\value{
the value when density function with parameter\code{(a,b)} at \code{x}.
}
\description{
The density function for Pareto distribution with parameter(a,b)
}
\examples{
\dontrun{
y <- seq(2,50,0.01)
plot(y, f_Pareto(y,2,2),col='blue'.type='l')
}
}
