% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Homework.R
\name{f_Sk}
\alias{f_Sk}
\title{The function need to be solved in Exercise 11.4, Homework7}
\usage{
f_Sk(k, a)
}
\arguments{
\item{k}{degrees of freedom}

\item{a}{the independent variables}
}
\value{
the difference between the two equations
}
\description{
To find the intersection points for S_k(a) and S_k-1(a), set f_S be the difference between the two equations and then get its root.
}
