% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Homework.R
\name{g_goal}
\alias{g_goal}
\title{The monte carlo in Exercise5.13, Homework3}
\usage{
g_goal(x)
}
\arguments{
\item{x}{the number of the samples you want to generate}
}
\value{
the value of density function at \code{x}
}
\description{
a special goal density we need to use the Monte Carlo method to integration
}
