% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Homework.R
\name{g_goal2}
\alias{g_goal2}
\title{The monte carlo in Exercise5.15, Homework3}
\usage{
g_goal2(x)
}
\arguments{
\item{x}{the value of the independent variable}
}
\value{
the value of the function at \code{x}
}
\description{
a special goal density we need to use the Monte Carlo piecewise method to integration
}
