% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Homework.R
\name{G_Rek}
\alias{G_Rek}
\title{rescaled Epanechnikov kernel in Exercise5.13, Homework3}
\usage{
G_Rek(n)
}
\arguments{
\item{n}{the number of the samples you want to generate}
}
\value{
the random samples followed the goal distribution
}
\description{
Devroye and Gyorfi's algorithm for simulation from the rescaled Epanechnikov kernel's distribution, in the Exercise 3.9, homework1.
}
\examples{
\dontrun{
x=G_Rek(1000)
hist(x, density = 3,breaks=30, col="red",prob = TRUE) #density histogram of sampl
}
}
