% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Homework.R
\name{Gelman.Rubin}
\alias{Gelman.Rubin}
\title{The Gelman-Rubin method to monitor convergence of the chain, Exercise 9.4+, Homework7}
\usage{
Gelman.Rubin(psi)
}
\arguments{
\item{psi}{psi[i,j] is the statistic psi(X[i,1:j])}
}
\value{
the estimated potential scale reduction
}
\description{
The Gelman-Rubin method of monitoring convergence of a M-H chain is based on comparing the behavior of several generated chains with respect to the variance of one or more scalar summary statistics. The estimates of the variance of the statistic are analogous to estimates based on between-sample and within-sample mean squared errors in a one-way analysis of variance (ANOVA)
}
