% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Research.R
\name{Generate.data}
\alias{Generate.data}
\title{Generate the sample data}
\usage{
Generate.data(Theta, Btensor, self = 0)
}
\arguments{
\item{Theta}{the matrix for theta(membership)}

\item{Btensor}{the community-wise connectivity parameter}

\item{self}{if self==0, the diagonal of the matrix is 1}
}
\value{
the multi-layer network sample
}
\description{
The function generate a \code{m*n*n} multi-layer network sample
}
\examples{
\dontrun{
n <- 200
k <- 3 # number of communities
m <- 3 # layers 
clust.size <- c(rep(floor(n/k), k-1),n-(k-1)*floor(n/k)) # equal cluster size
Theta <- Generate.theta(n, k, clust.size)
Btensor <- array(0, dim = c(3, 3, 3)) # B tensor should be m*k*k
Btensor[1, ,] <- matrix(c(0.6,0.4,0.4,0.4,0.2,0.2,0.4,0.2,0.2), 3, 3)
Btensor[2, ,] <- matrix(c(0.2,0.4,0.2,0.4,0.6,0.4,0.2,0.4,0.2), 3, 3)
Btensor[3, ,] <- matrix(c(0.2,0.2,0.4,0.2,0.2,0.4,0.4,0.4,0.6), 3, 3)
A <- Generate.data(Theta, Btensor, self = 0)
}
}
