% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Research.R
\name{Generate.theta}
\alias{Generate.theta}
\title{Generate the  theta}
\usage{
Generate.theta(n, K, clust.size)
}
\arguments{
\item{n}{the number of the nodes}

\item{K}{the number of the communities}

\item{clust.size}{the vector, each element represents the number of a taxon}
}
\value{
the matrix for theta
}
\description{
The function generate a matrix for the theta, which will be used in generate data
}
\examples{
\dontrun{
n <- 200
k <- 3 # number of communities
m <- 3 # layers 
clust.size <- c(rep(floor(n/k), k-1),n-(k-1)*floor(n/k)) # equal cluster size
Theta <- Generate.theta(n, k, clust.size)
}
}
