% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Research.R
\name{GetCenter}
\alias{GetCenter}
\title{Get the initial estimated Btensor}
\usage{
GetCenter(A, idx, k)
}
\arguments{
\item{A}{the sample m*n*n multi-layer network}

\item{idx}{the initial estiamted membership}

\item{k}{the number of the communities}
}
\value{
the initial estimated Btensor
}
\description{
The function obtain an initial estimator of Btensor by simple processing of sample A
}
