% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetRandomNumber}
\alias{GetRandomNumber}
\title{Runif using Rcpp, in Ex2, Homework9}
\usage{
GetRandomNumber(N)
}
\arguments{
\item{N}{the number of the sample}
}
\value{
a vector of random sample
}
\description{
The function is for randomly generating sample that are uniformly distributed on (0,1) with Rcpp
}
