% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Homework.R
\name{Laplace.chain}
\alias{Laplace.chain}
\title{The Metropolis chain in Exercise 9.4+, Homework7}
\usage{
Laplace.chain(sigma, N, X1, f)
}
\arguments{
\item{sigma}{the standard deviation of a normal distribution}

\item{N}{length of chains}

\item{X1}{the initial value of the chains}

\item{f}{the density function of standard Laplace distribution}
}
\value{
the chains
}
\description{
Generates a Metropolis chain for Laplace distribution with Normal(X[t], sigma) proposal distribution and starting value X1
}
