% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Homework.R
\name{maxout}
\alias{maxout}
\title{Maximum number of extreme points in Exercises 8.3,Homework6}
\usage{
maxout(x, y)
}
\arguments{
\item{x}{the sample X}

\item{y}{the sample y}
}
\value{
the maximum number of extreme points
}
\description{
The function calculate the maximum number of extreme points that applies when sample sizes are not necessarily equal
}
\examples{
\dontrun{
X=rnorm(1000,0,2)
Y=rnorm(1000,1,1)
maxout(X,Y)
}
}
