% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Research.R
\name{Purity}
\alias{Purity}
\title{Purity-An external measures}
\usage{
Purity(f1, f2, K, n)
}
\arguments{
\item{f1}{the real classify label}

\item{f2}{the estimated cluster label}

\item{K}{the number of the communities}

\item{n}{the number of the nodes}
}
\value{
the proportion of the classified successfully nodes
}
\description{
Purity is a simple and transparent evaluation measure, when the value of purity is close to 1, tell the result of cluster is very good.
}
\examples{
\dontrun{
y <- seq(2,50,0.01)
plot(y, f(y,2,2),col='blue'.type='l')
}
}
