% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Homework.R
\name{rw.Metropolis}
\alias{rw.Metropolis}
\title{A random walk Metropolis sampler in Exercise 9.4, Homework7}
\usage{
rw.Metropolis(sigma, x0, N, f)
}
\arguments{
\item{sigma}{the standard deviation of a normal distribution}

\item{x0}{initial value}

\item{N}{the number of the goal sample}

\item{f}{the density function of the sample}
}
\value{
the list of sample subject to the goal distribution
}
\description{
This is a random walk Metropolis sampler for generating the standard Laplace distribution (see Exercise 3.2). For the increment, simulate from a normal distribution.
}
