% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rwMC}
\alias{rwMC}
\title{rw.Metropolis using Rcpp, in Ex1, Homework9}
\usage{
rwMC(sigma, x0, N)
}
\arguments{
\item{sigma}{the standard deviation of the normal distribution}

\item{x0}{the initial value of the chains}

\item{N}{the length of the chains}
}
\value{
a matrix,include the value of each step, and the reject times
}
\description{
The function is for random walk Metropolis sampler to generate the standard Laplace distribution, which is written in Rcpp
}
