% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Homework.R
\name{sk}
\alias{sk}
\title{Estimate the power of the skewness test in Exercise 6.7 in Homework4}
\usage{
sk(x)
}
\arguments{
\item{x}{the sample}
}
\value{
the value of the sample \code{x}'s skewness
}
\description{
The function is used to computes the sample's skewness coeff
}
\examples{
 
\dontrun{
X=rnorm(1000,0,2)
sk(X)
}
}
