% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Homework.R
\name{sk_m}
\alias{sk_m}
\title{Estimate the power of Mardia's multivariate skewness test for in Exercise 6.C in Homework4}
\usage{
sk_m(X)
}
\arguments{
\item{X}{the multivariate sample}
}
\value{
the value of the multivariate skewness statistic
}
\description{
Mardia proposed tests of multivariate normality based on multivariate generalizations of skewness and kurtosis.
}
