\name{StatComp20025-package}
\alias{StatComp20025-package}
\alias{StatComp20025}
\docType{package}
\title{The final job of Statistical Calculation in 2020 in USTC }
\description{
This package includes functions used for the fall 2020 course assignments and for research.
}
\details{
 Some functions in this package, such as Generate.Theta, Generate.Data, GetCenter, GetDist, are very useful in my research. The rest of the functions are used in my homework.
}
\author{
Hanyang Tian

Maintainer: THY <thy@mail.ustc.edu.cn>
}
\references{
  The functions used in my research mainly should be for the Jing Lei's article,"Consistent community detection in multi-layer network data". And the functions used in my homework mainly refer to the book, "Statistical Computing with R".
}
\keyword{ package }


