% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Homework.R
\name{Tn}
\alias{Tn}
\title{Nearest neighbor tests in Discussion, Homework6}
\usage{
Tn(z, ix, sizes, k)
}
\arguments{
\item{z}{the matrix of sample X and Y}

\item{ix}{the index of sample}

\item{sizes}{[1]the number of sample x; [2] the number of sample y}

\item{k}{the nearest neighbor statistic measures the proportion of first through \code{k}th nearest neighbor coincidences}
}
\value{
the proportion of first through \code{k}th nearest neighbor coincidences
}
\description{
The nearest neighbor (NN) tests are a type of test based on ordered distances between sample elements, which can be applied when the distributions are continuous.
}
